requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.9
django-cors-headers==3.14.0
django-environ==0.9.0
django-filter==23.1
django-storages==1.13
djangorestframework==3.14.0
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-core==2.3.2
google-cloud-secret-manager==2.12.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.48.2
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
Pillow==9.5.0
proto-plus==1.22.2
protobuf==3.20.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
